Revert "m_option: change m_option_type_aspect to double"
This reverts commit 9eb53dd4f3598c1e40c13bf7e6c502b3393afa04.

+++ filters/f_decoder_wrapper.c
@@ -93,7 +93,7 @@ static const struct m_sub_options adec_queue_conf = {
 #define OPT_BASE_STRUCT struct dec_wrapper_opts
 
 struct dec_wrapper_opts {
-    double movie_aspect;
+    float movie_aspect;
     int aspect_method;
     double force_fps;
     bool correct_pts;
